requirements.txt
streamlit
yfinance
twilio
python-dotenv
google-generativeai
openai>=1.0.0
plotly
schedule
pyyaml>=6.0
tqdm>=4.65.0

# API Dependencies
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
jinja2>=3.1.0
python-multipart>=0.0.6

# Authentication
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4

# ML Forecasting Dependencies
torch>=2.0.0
transformers>=4.35.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.3.0

# Signal Extraction Dependencies
fredapi>=0.5.0
feedparser>=6.0.10
